feat: Add Model Metabox
