+++ trunk/src/Metabox/Type/class-post.php
@@ -3,19 +3,22 @@
 namespace Trasweb\Plugins\DisplayMetadata\Metabox\Type;
 
 use Trasweb\Plugins\DisplayMetadata\Metabox\Metabox;
-use Trasweb\Plugins\DisplayMetadata\Plugin;
-use const ARRAY_A;
+use Trasweb\Plugins\DisplayMetadata\Metabox\Model;
 
 /**
  * This class manages `Display Metadata` post metabox.
  */
 final class Post extends Metabox {
-	protected const TITLE       = 'Post information';
-
-	/**
-	 * @var string Field name where meta is saved for item_id
-	 */
-	protected const FIELD_META_ID = 'post_id';
+    /**
+     * Metabox constructor
+     *
+     * @return void
+     */
+    public function __construct(int $item_id = 0, ?Model $model = null)
+    {
+        $this->item_id = $item_id;
+        $this->model = $model ?? new Model\Post_Model($item_id);
+    }
 
 	/**
 	 * Register a metabox in order to display it later.
@@ -25,7 +28,7 @@ final class Post extends Metabox {
 	public function register(): void
 	{
 		add_meta_box( 'trasweb_metadata_metabox',     // Unique ID
-			static::TITLE,                            // Box title
+            $this->model->get_title(),                 // Box title
 			[ $this, 'display' ],                     // Content callback
 			$this->get_accepted_cpt()          // Post type
 		);
@@ -53,27 +56,6 @@ final class Post extends Metabox {
 		return array_values( get_post_types( [ 'show_ui' => true ] ) ?: [] );
 	}
 
-	/**
-	 * Retrieve item properties/fields. E.g: ID, post_title, ...
-	 *
-	 * @return array
-	 */
-	public function get_item_properties(): array
-	{
-		return get_post( $this->item_id, ARRAY_A ) ?: [];
-	}
-
-	/**
-	 * Retrieve metadata table name for current WordPress
-	 *
-	 * @return string table name.
-	 */
-	protected function get_meta_table_name(): string {
-		global $wpdb;
-
-		return $wpdb->postmeta;
-	}
-
     /**
      * Display metadata metabox.
      *
@@ -81,6 +63,6 @@ final class Post extends Metabox {
      */
     public function display(): void
     {
-        do_action('trasweb_metabox_display', $this);
+        do_action('trasweb_metabox_display', $this->get_model());
     }
 }
