+++ trunk/src/Metabox/Type/class-user.php
@@ -4,17 +4,22 @@ namespace Trasweb\Plugins\DisplayMetadata\Metabox\Type;
 
 use Trasweb\Plugins\DisplayMetadata\Metabox\Metabox;
 
+use Trasweb\Plugins\DisplayMetadata\Metabox\Model;
+
 /**
  * This class manages `Display Metadata` user metabox.
  */
 final class User extends Metabox {
-
-	protected const TITLE = 'User information';
-
-	/**
-	 * @var string Field name where meta is saved for item_id
-	 */
-    public const FIELD_META_ID = 'user_id';
+    /**
+     * Metabox constructor
+     *
+     * @return void
+     */
+    public function __construct(int $item_id = 0, ?Model $model = null)
+    {
+        $this->item_id = $item_id;
+        $this->model = $model ?? new Model\User_Model($item_id);
+    }
 
 	/**
 	 * Register a metabox in order to display it later.
@@ -38,23 +43,13 @@ final class User extends Metabox {
         return 'user-edit' === $screen_slug || 'profile' === $screen_slug;
     }
 
-	/**
-	 * Retrieve item properties/fields. E.g: ID, user_login, ...
-	 *
-	 * @return array
-	 */
-	public function get_item_properties(): array {
-		return \json_decode( \json_encode( get_user_by('id', $this->item_id ) ), true );
-	}
-
-	/**
-	 * Retrieve metadata table name for current WordPress
-	 *
-	 * @return string table name.
-	 */
-	protected function get_meta_table_name(): string {
-		global $wpdb;
-
-		return $wpdb->usermeta;
-	}
+    /**
+     * Return a model for current Metabox
+     *
+     * @return Model
+     */
+    public function get_model(): Model
+    {
+        return new Model\User_Model($this->item_id);
+    }
 }
