+++ trunk/src/Metabox/class-metabox-view.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Trasweb\Plugins\DisplayMetadata\Metabox;
 
 use Trasweb\Plugins\DisplayMetadata\Plugin;
+use Trasweb\Plugins\DisplayMetadata\Metabox\Model;
 
 use const Trasweb\Plugins\DisplayMetadata\PLUGIN_NAME;
 
@@ -14,21 +15,21 @@ class Metabox_View
 {
     public const VIEWS_PATH = Plugin::PATH . '/../views';
 
-    private Metabox $metabox;
+    private Model $metabox_model;
 
     /**
-     * @param Metabox $metabox
+     * @param Model $metabox_model
      * @return void
      */
-    public function __construct( Metabox $metabox) {
-        $this->metabox = $metabox;
+    public function __construct( Model $metabox_model) {
+        $this->metabox_model = $metabox_model;
     }
 
 
     public function display(string $metabox_type = 'metabox', string $metabox_path = self::VIEWS_PATH) {
         $metabox_type = sanitize_key($metabox_type);
 
-        $metabox_title = $this->metabox->get_title();
+        $metabox_title = $this->metabox_model->get_title();
         $metadata_list = $this->get_metadata_list();
 
         include $metabox_path."/{$metabox_type}.php";
@@ -40,8 +41,8 @@ class Metabox_View
      * @return Metadata_Iterator
      */
     public function get_metadata_list(): Metadata_Iterator {
-        $item_properties = $this->metabox->get_item_properties();
-        $item_metadata = $this->metabox->get_item_metadata();
+        $item_properties = $this->metabox_model->get_item_properties();
+        $item_metadata = $this->metabox_model->get_item_metadata();
 
         $item_vars = [
             __( 'Properties', PLUGIN_NAME ) => $item_properties,
